requirements.txt
bs4
openai
pandas
python-dotenv
requests
numpy
beautifulsoup4
python-dotenv
lxml
bs4
openai
pandas
python-dotenv
requests
numpy
beautifulsoup4
python-dotenv
lxml